Rename matomo directory to matomo-config
This prevents adblockers like uBlock Origin from blocking our setup scripts.
The adblockers will still block matomo.php, but our source code will not be affected.
