+++ projects/bp-gallery/src/helpers/history.ts
@@ -1,4 +1,4 @@
-import { trackHistory } from '../matomo/matomo';
+import { trackHistory } from '../matomo-config/matomo';
 
 export const pushHistoryWithoutRouter = (newLocation: string) => {
   window.history.pushState({}, '', newLocation);
