+++ projects/bp-gallery/src/matomo-config/TrackHistoryWithMatomo.tsx
