+++ projects/bp-gallery/src/matomo-config/matomo.ts
